+++ 14.STM32F401RET6-FreeRTOS/HARDWARE/inc/w25q64.h
@@ -13,6 +13,6 @@ u8 w25q64_read_busy(void);
 void w25q64_sector_erase(u32 addr);
 void w25q64_write_page(u32 addr,u8 data_size,u8 *data);
 void w25q64_cross_page_write(u32 addr,u16 data_size,u8 *data);
-void w25q64_read_data(u32 addr,u8 data_size,u8 *data);
+void w25q64_read_data(u32 addr,u16 data_size,u8 *data);
 void w25q64_block_erase(u32 addr);
 #endif
